Fucking casts. dabbit

+++ Contrary/Contrary/Constants.h
@@ -5,4 +5,4 @@ extern int SCREEN_H;
 extern int SCREEN_W;
 extern int TICK;
 extern double TIME_SCALE;
-extern int FPS;
+extern double FPS;
